refactor(renderer): split content stats and crop helpers
Move raster content measurement and viewport crop conversion out of the renderer coordinator while preserving the existing renderer module exports for callers.
